角色连表查询权限

+++ DatabaseProject/HR_java/src/main/java/com/hr_java/controller/RoleController.java
@@ -27,7 +27,7 @@ public class RoleController {
     }
 
     @RequestMapping("/{id}")
-    public Result get(@PathVariable("id") Long id) {
+    public Result get(@PathVariable("id") Integer id) {
         return Result.succ(roleService.getById(id));
     }
 }
